+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/channel/dto/response/ChannelSummaryDto.java
@@ -2,7 +2,7 @@ package com.hertz.hertz_be.domain.channel.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.hertz.hertz_be.domain.channel.repository.projection.ChannelRoomProjection;
-import com.hertz.hertz_be.global.common.AESUtil;
+import com.hertz.hertz_be.global.util.AESUtil;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
