+++ hertz-be/src/main/java/com/hertz/hertz_be/global/util/AESUtil.java
@@ -1,4 +1,4 @@
-package com.hertz.hertz_be.global.common;
+package com.hertz.hertz_be.global.util;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
